# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.RelationRequest:
    properties:
      from_entity_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      from_entity_type:
        example: user
        type: string
      properties:
        type: object
      relation_type:
        example: owns
        type: string
      to_entity_id:
        example: 650e8400-e29b-41d4-a716-446655440001
        type: string
      to_entity_type:
        example: document
        type: string
    required:
    - from_entity_id
    - from_entity_type
    - relation_type
    - to_entity_id
    - to_entity_type
    type: object
  handlers.RelationResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      from_entity_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      from_entity_type:
        example: user
        type: string
      id:
        example: 750e8400-e29b-41d4-a716-446655440002
        type: string
      properties:
        type: object
      relation_type:
        example: owns
        type: string
      to_entity_id:
        example: 650e8400-e29b-41d4-a716-446655440001
        type: string
      to_entity_type:
        example: document
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  middleware.ErrorDetail:
    description: Detailed error information including code, message, and optional
      details
    properties:
      code:
        example: NOT_FOUND
        type: string
      details:
        type: object
      message:
        example: Entity not found
        type: string
      request_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  middleware.ErrorResponse:
    description: Standard error response format for all API errors
    properties:
      error:
        $ref: '#/definitions/middleware.ErrorDetail'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@entropic.io
    name: API Support
    url: http://www.entropic.io/support
  description: A next-generation storage engine with flexible entity-relationship
    model and dual-storage architecture
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Entropic Storage Engine API
  version: "1.0"
paths:
...
  /api/v1/relations:
    post:
      consumes:
      - application/json
      description: Create a new relation between two entities
      parameters:
      - description: Relation data
        in: body
        name: relation
        required: true
        schema:
          $ref: '#/definitions/handlers.RelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.RelationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Create a new relation
      tags:
      - relations
  /api/v1/relations/{relationID}:
    delete:
      consumes:
      - application/json
      description: Delete a specific relation by its ID
      parameters:
      - description: Relation ID
        format: uuid
        in: path
        name: relationID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Delete a relation
      tags:
      - relations
    get:
      consumes:
      - application/json
      description: Get a specific relation by its ID
      parameters:
      - description: Relation ID
        format: uuid
        in: path
        name: relationID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RelationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middleware.ErrorResponse'
      summary: Get a relation by ID
      tags:
      - relations
...
schemes:
- http
- https
swagger: "2.0"